% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_mirtCAT_computeQuestion_05-06-19.R
\name{computeQuestion}
\alias{computeQuestion}
\title{computeQuestion}
\usage{
computeQuestion(CATdesign, askedQuestions, responses)
}
\arguments{
\item{CATdesign}{mirtCAT design object, obtained using mirtCAT(), see example}

\item{askedQuestions}{The ID's of the questions administered so far}

\item{responses}{The responses to each questions administered so far}
}
\value{
nextQuestion: the ID of the next question to administer or "done", indicating the termination of CAT
}
\description{
Find the question ID for the next question to be administered in CAT
}
\examples{
# Specify CATdesign object
data(CATdesign)
CATdesign <- mirtCAT(df, mod, criteria = 'KL', start_item = 'Trule',
                     design_elements = TRUE,
                     design = list(min_SEM = rep(0.4, 3),
                                   max_items = ncol(data_epsi1a),
                                   delta_thetas = rep(0.03, 3)))
# Questions #1 and #5 were previously administered
askedQuestions = c(1,5)
# The respondent's response to the previously administered questions (i.e., Questions #1 and #5) were Options 1 and 2
responses = c(1,2)
nextQuestion <- computeQuestion(CATdesign = CATdesign, askedQuestions = askedQuestions, responses = responses)
# The ID of the next question to administer
nextQuestion

}
